+++ Object.h
@@ -24,7 +24,7 @@ public:
 	virtual ~Object();
 
 	/*Given a ray, this method calculates the intersection with the object and returns the position of the intersection*/
-	virtual float computeIntersection(Ray* ray);
+	virtual double computeIntersection(Ray* ray);
 
 	/*Given a position, this method calculates de normal vector at that point*/
 	virtual Vec3d* computeNormal(Vec3d* position);
